+++ docs/history.md
@@ -3,8 +3,10 @@
 <!-- cspell:ignore antizoubilamaka,Erbear,Neonomi,slvn -->
 <!-- markdownlint-disable MD001 -->
 
-### _v0.2.4_ - October 15th, 2023 (Unreleased)
+### _v0.2.4_ - October 18th, 2023 (Unreleased)
 
+- In hardcore mode, pill effects will now unlock on a cycle of one negative, one neutral, and one positive.
+- Rooms now have a chance to be flipped on the X axis, Y axis, or both axes.
 - A red icon will now show if you are playing on hardcore mode.
 - The mod will now display a message if you illegally pause the game or save & quit.
 - The stats page will now show if you have illegally paused or saved & quit.